requirements.txt
streamlit
pandas
plotly.express
altair
matplotlib.pyplot